requirements.txt
kafka-python
ollama
requests